Create Plus_Minus.cpp
Given an array of integers, calculate the fractions of its elements that are positive, negative, and are zeros. Print the decimal value of each fraction on a new line.
Input : enter the size of array and the elements . 
output : the ratio of positive , negative and zeros in the array respectively.